Extract accept core into shared leaf module (task #3)
Move the acceptance body out of the accept_match_result router handler into
api/app/result_acceptance.py (accept_standing_result + the rating/side-won
helpers) so a worker can reuse it without importing router internals. The core
signals a superseded/already-accepted head via a domain result instead of
raising HTTPException; the router maps that to the existing 409. No behavior
change — 41 acceptance/result/negotiation tests green. Prereq for the
retirement job (task #4).
